# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rawand Decoration CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workplaces:
    get:
      operationId: listWorkplaces
      tags: [organization]
      summary: List active workplaces
      responses:
        "200":
          description: Workplace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workplace"
    post:
      operationId: createWorkplace
      tags: [organization]
      summary: Create a workplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkplaceInput"
      responses:
        "201":
          description: Workplace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workplace"
  /groups:
    get:
      operationId: listGroups
      tags: [organization]
      parameters:
        - $ref: "#/components/parameters/WorkplaceIdQuery"
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [organization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /users:
    get:
      operationId: listUsers
      tags: [organization]
      parameters:
        - $ref: "#/components/parameters/WorkplaceIdQuery"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [organization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /employees:
    get:
      operationId: listEmployees
      tags: [organization]
      parameters:
        - $ref: "#/components/parameters/WorkplaceIdQuery"
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [organization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /drivers:
    get:
      operationId: listDrivers
      tags: [organization]
      parameters:
        - $ref: "#/components/parameters/WorkplaceIdQuery"
      responses:
        "200":
          description: Driver list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [organization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Driver created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
...
components:
  parameters:
...
    WorkplaceIdQuery:
      in: query
      name: workplaceId
      schema:
        type: number
        minimum: 1
  schemas:
...
    Workplace:
      type: object
      required: [id, name, code, address, phone, email, currency, status, createdAt]
      properties:
        id: { type: number }
        name: { type: string }
        code: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        currency: { type: string }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
    WorkplaceInput:
      type: object
      required: [name, code, currency]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        address: { type: string, default: "" }
        phone: { type: string, default: "" }
        email: { type: string, default: "" }
        currency: { type: string, minLength: 1 }
    Group:
      type: object
      required: [id, name, permissions, status, createdAt]
      properties:
        id: { type: number }
        workplaceId: { type: [number, "null"] }
        name: { type: string }
        permissions:
          type: array
          items: { type: string }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
    GroupInput:
      type: object
      required: [name]
      properties:
        workplaceId: { type: [number, "null"] }
        name: { type: string, minLength: 1 }
        permissions:
          type: array
          items: { type: string }
          default: []
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id: { type: number }
        workplaceId: { type: [number, "null"] }
        groupId: { type: [number, "null"] }
        username: { type: string }
        displayName: { type: string }
        status: { type: string, enum: [active, inactive] }
        lastLoginAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [username, displayName, password]
      properties:
        workplaceId: { type: [number, "null"] }
        groupId: { type: [number, "null"] }
        username: { type: string, minLength: 1 }
        displayName: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
    Employee:
      type: object
      required: [id, workplaceId, name, phone, email, address, jobTitle, department, status, createdAt]
      properties:
        id: { type: number }
        workplaceId: { type: number }
        userId: { type: [number, "null"] }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        jobTitle: { type: string }
        department: { type: string }
        startDate: { type: [string, "null"], format: date }
        endDate: { type: [string, "null"], format: date }
        compensation: { type: [number, "null"] }
        compensationCurrency: { type: [string, "null"] }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
    EmployeeInput:
      type: object
      required: [workplaceId, name]
      properties:
        workplaceId: { type: number, minimum: 1 }
        userId: { type: [number, "null"] }
        name: { type: string, minLength: 1 }
        phone: { type: string, default: "" }
        email: { type: string, default: "" }
        address: { type: string, default: "" }
        jobTitle: { type: string, default: "" }
        department: { type: string, default: "" }
        startDate: { type: [string, "null"], format: date }
        endDate: { type: [string, "null"], format: date }
        compensation: { type: [number, "null"], minimum: 0 }
        compensationCurrency: { type: [string, "null"] }
    Driver:
      type: object
      required: [id, workplaceId, name, phone, vehicle, licenseNumber, isAssigned, status, createdAt]
      properties:
        id: { type: number }
        workplaceId: { type: number }
        employeeId: { type: [number, "null"] }
        name: { type: string }
        phone: { type: string }
        vehicle: { type: string }
        licenseNumber: { type: string }
        isAssigned: { type: boolean }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
    DriverInput:
      type: object
      required: [workplaceId, name]
      properties:
        workplaceId: { type: number, minimum: 1 }
        employeeId: { type: [number, "null"] }
        name: { type: string, minLength: 1 }
        phone: { type: string, default: "" }
        vehicle: { type: string, default: "" }
        licenseNumber: { type: string, default: "" }
        isAssigned: { type: boolean, default: false }
...