# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rawand Decoration CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: number;
  workplaceId: number;
  /** @nullable */
  userId?: number | null;
  name: string;
  phone: string;
  email: string;
  address: string;
  jobTitle: string;
  department: string;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  endDate?: Date | null;
  /** @nullable */
  compensation?: number | null;
  /** @nullable */
  compensationCurrency?: string | null;
  status: EmployeeStatus;
  createdAt: Date;
}