# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rawand Decoration CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  /** @minimum 1 */
  workplaceId: number;
  /** @nullable */
  userId?: number | null;
  /** @minLength 1 */
  name: string;
  phone?: string;
  email?: string;
  address?: string;
  jobTitle?: string;
  department?: string;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  endDate?: Date | null;
  /**
     * @minimum 0
     * @nullable
     */
  compensation?: number | null;
  /** @nullable */
  compensationCurrency?: string | null;
}