# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rawand Decoration CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  /** @nullable */
  workplaceId?: number | null;
  /** @minLength 1 */
  name: string;
  permissions?: string[];
}