# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rawand Decoration CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  /** @nullable */
  workplaceId?: number | null;
  /** @nullable */
  groupId?: number | null;
  username: string;
  displayName: string;
  status: UserStatus;
  /** @nullable */
  lastLoginAt?: Date | null;
  createdAt: Date;
}