# lib/api-zod/src/generated/types/workplace.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rawand Decoration CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkplaceStatus } from './workplaceStatus';

export interface Workplace {
  id: number;
  name: string;
  code: string;
  address: string;
  phone: string;
  email: string;
  currency: string;
  status: WorkplaceStatus;
  createdAt: Date;
}