# lib/api-zod/src/generated/types/workplaceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rawand Decoration CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkplaceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  address?: string;
  phone?: string;
  email?: string;
  /** @minLength 1 */
  currency: string;
}